+++ docs/RELEASE-NOTES.md
@@ -1,3 +1,4 @@
 ### Fixed
 
 - Images not showing up for users with an `'` in their file path.
+- Issue with dimension capping of assets.